projects/django/ominousshell/requirements.txt
mysql-connector-python
mysqlclient
Django
